Fix the Windows build and linter definitions
